fix: RECAP-1, STUCKCONN-1, PACETYPE-1 — three real bugs, all verified live
Three separate, previously-undiagnosed defects found and fixed while
investigating "everything is still wrong" — traced with real production
data on the simulator, not theorized, per Rule 13.

RECAP-1 · a day the runner has already RUN was showing its authored plan
instead of what actually happened. David, live: "this is not post run
display... its not showing what I did" — true of EVERY past run, not
just one. `PlanSnapshotDay` carries authored structure only, by design;
the fix is not to teach it a second recap renderer but to stop routing a
day with a `matched_run` through it at all. `shouldRenderFromSnapshot`
(and the matching guard in `TodayHostV5.body`'s snapshot branch) now
excludes any day with a matched run, so it falls through to the SAME live
`/api/v5/today?date=` fetch `isHome` already uses — verified directly:
`date=2026-09-03` returns `state: "after_run"` with the full recap
payload (splits, actual pace, HR zones, route, coach's read), and
`TodayAfterV5` renders it correctly, because that machinery already
existed and worked — it was simply never reached for a browsed date.

STUCKCONN-1 · reproduced live in the simulator: every single endpoint
(today, races, block, plan-snapshot, watch) timing out identically,
while a plain `curl` to the same production host from the same machine
succeeded in under a second throughout. The OS log named the actual
cause: "HTTP/2 terminating broken Connection 3, last success 867595s
ago, consecutive failures 11" — a connection `URLSession.shared` had
been silently reusing and re-failing on for roughly ten days. This was
never a redeploy or a server issue; "Can't reach faff" was completely
honest about not reaching the server, just not about why. Added
`StuckConnectionMonitor`: after 3 consecutive `.timedOut` failures across
independent `authedSend` calls, fires exactly one
`URLSession.shared.reset(completionHandler:)` to force fresh connections,
never on a single blip and never while a reset is already in flight.

PACETYPE-1 · `V5SupplementalRun.paceSPerMi` was declared `Int?`; the
server's own wire type is `number | null` and the value it actually sends
(elapsed seconds / distance) is essentially never a whole number.
`APIRow.list(_:)` decodes an array with `try?` and returns `[]` on ANY
element's decode failure — Rule 11: a swallowed failure, not a genuine
empty read — so a fractional pace on one supplemental run silently erased
the WHOLE array, every time, for every runner. David: "not loading in my
treadmill run or the extra easy run" — the treadmill run (matched,
prescribed) was RECAP-1's bug; the extra easy run (supplemental) was
this one, independently. Widened to `Double?`, matching the server
contract and `PlanSnapshotModels.swift`'s sibling structs, which already
had this right.

All three verified against David's real account and real data in the
simulator, not a fixture: Wednesday 9/3 now shows the treadmill
INTERVALS recap (4.71 mi actual, not the plan's 6.0 mi) AND "ALSO TODAY /
4.5 easy / 8:19" for the supplemental run, both on a clean load with no
stale banner.

+++ native-v2/Faff/Faff/DesignV5/APIV5.swift
@@ -519,12 +519,26 @@ struct V5OffSeason: Decodable, Equatable {
 /// `lib/faff/v5-today.ts`'s `V5SupplementalRunWire`. Deliberately lean, no
 /// verdict and no workout type: a supplemental run was never shown to
 /// execute anything prescribed, so nothing here may imply that it did.
+///
+/// PACETYPE-1 (2026-09-04) · `paceSPerMi` was `Int?`, but the server's own
+/// wire type is `number | null` (`V5SupplementalRunWire`) and the value it
+/// actually sends — `runFacts(...).paceSecPerMi`, elapsed seconds divided by
+/// distance — is essentially never a whole number. `APIRow.list(_:)`
+/// decodes an array with `try?` and returns `[]` on ANY element failure
+/// (Rule 11: a swallowed decode failure, not a genuine empty read), so a
+/// fractional pace on ONE supplemental run silently erased the WHOLE array
+/// for that day, every day, unconditionally. Caught live: David, on a day
+/// with a real second run logged, "not loading in my treadmill run or the
+/// extra easy run" — the treadmill run (the matched, prescribed one) was
+/// never affected by this, only the supplemental one, because only
+/// `V5SupplementalRun` carried the wrong type; `PlanSnapshotModels.swift`'s
+/// sibling structs already had this right as `Double?`.
 struct V5SupplementalRun: Decodable, Equatable, Identifiable {
     var id: String { runId }
     let runId: String
     let distanceMi: Double
     let durationSec: Int?
-    let paceSPerMi: Int?
+    let paceSPerMi: Double?
     let indoor: Bool
 }
